ai-atari/gcp_deep_learning_vm/trainVM/requirements(on_deep_learning_VM_GCP).txt
#This is pip freeze on deep learning VM environment offered by GCP (VM comes pre-installed with GPU drivers & tensorflow)
absl-py==0.9.0
astor==0.8.1
atari-py==0.2.6
box2d-py==2.3.8
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cloudpickle==1.3.0
cycler==0.10.0
Cython==0.29.15
future==0.18.2
gast==0.2.2
glfw==1.11.0
google-auth==1.11.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
gym==0.17.1
h5py==2.10.0
idna==2.9
imageio==2.8.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lockfile==0.12.2
Markdown==3.2.1
#matplotlib==3.2.1
#mock==4.0.2
numpy==1.18.2
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.2.0
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyglet==1.5.0
pyparsing==2.4.6
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
###tensorboard==2.0
#tensorflow==1.13.1
#tensorflow-estimator==1.13.0
termcolor==1.1.0
tqdm==4.43.0
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.12.1
#tensorboard==2.1.0
#tensorflow-datasets==1.3.0
#tensorflow-estimator==2.1.0
#tensorflow-gpu==2.1.0
#tensorflow-hub==0.7.0
#tensorflow-io==0.9.10
#tensorflow-metadata==0.21.1
#tensorflow-probability==0.9.0
#tensorflow-serving-api-gpu==1.14.0

